+++ ecorise-react-app/src/Reward/style.css
@@ -6,12 +6,12 @@ html, body {
 }
 
 .main-content {
-  padding: 20px;
+  padding: 20px ;
   font-family: Arial, sans-serif;
   background: white;
   min-height: 100vh;
   max-width: 1200px;
-  margin: 0 auto;
+  margin-left: 22%;
 }
 
 .rewards-title {
@@ -30,6 +30,7 @@ html, body {
   border-radius: 12px;
   box-shadow: 0 4px 15px rgba(0,0,0,0.07);
   overflow-x: auto;
+  margin: 0 auto;
 }
