+++ ecorise-react-app/src/utils/fetchRewards.js
@@ -1,5 +1,5 @@
 
-export async function fetchJSON(url, options = {}) {
+export async function fetchReward(url, options = {}) {
   const response = await fetch(url, options);
   if (!response.ok) {
     throw new Error('Network response was not ok');
